Move AtsError from TypeSystem into RemoteHost
It is not need publicly, it is just used to remote exception info.
