+++ src/Aspire.Hosting.RemoteHost/Ats/CapabilityException.cs
@@ -1,8 +1,6 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Aspire.TypeSystem;
-
 namespace Aspire.Hosting.RemoteHost.Ats;
 
 /// <summary>
