+++ tests/Aspire.Hosting.RemoteHost.Tests/CapabilityDispatcherTests.cs
@@ -3,8 +3,8 @@
 
 using System.Text.Json.Nodes;
 using Aspire.Hosting.ApplicationModel;
-using Aspire.TypeSystem;
 using Aspire.Hosting.RemoteHost.Ats;
+using Aspire.TypeSystem;
 using Xunit;
 
 namespace Aspire.Hosting.RemoteHost.Tests;
