+++ tests/Aspire.Hosting.RemoteHost.Tests/ReferenceExpressionRefTests.cs
@@ -3,7 +3,6 @@
 
 using System.Text.Json.Nodes;
 using Aspire.Hosting.ApplicationModel;
-using Aspire.TypeSystem;
 using Aspire.Hosting.RemoteHost.Ats;
 using Xunit;
